Adds missing cast
This fixes:
warning C4244: 'return': conversion from '__int64' to 'uint32_t', possible loss of data

+++ include/tao/json/external/ryu.hpp
@@ -1025,7 +1025,7 @@ namespace tao
             if( exp ) {
                result[ index++ ] = 'e';
                const auto end = jeaiii::i32toa( exp, result + index );
-               return end - result;
+               return static_cast<std::uint32_t>( end - result );
             }
             return index;
          }
